+++ hass-actronque/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [v0.27] - 2021-08-30
+
+### Added
+- The outside temperature reading from the Master Controller is now presented as an entity.
+- The compressor capacity reading from the Master Controller is now presented as an entity.
+- The zone position reading from each zone is now presented as an entity.
+
+### Changed
+- Zones that are off will now receive set temperature updates.
+
 ## [v0.26] - 2021-08-01
 
 ### Updated